更新数据类型为 float64，优化重构函数中的数值稳定性，调整数据加载函数以支持新的参数，删除不再使用的 generate.py 文件，更新 requirements.txt 以反映新依赖

+++ train.py
@@ -60,7 +60,7 @@ def train(is_load=False):
 
     for epoch in range(NUM_EPOCHS):
         for i, [data] in enumerate(train_loader):
-            data = data.to(DEVICE, dtype=torch.float32).view(data.shape[0], INPUT_DIM)
+            data = data.to(DEVICE, dtype=torch.float64).view(data.shape[0], INPUT_DIM)
             x_reconstructed, mu, log_var = model(data)
 
             # Compute loss
